Move core to packages
