+++ packages/core/generator.js
