+++ packages/core/index.js
@@ -1,0 +1,8 @@
+module.exports = {
+  ...require("./compiler"),
+  ...require("./generator"),
+  ...require("./lexer"),
+  ...require("./model"),
+  ...require("./parser"),
+  ...require("./patterns")
+};
