+++ packages/core/lexer.js
