+++ packages/core/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "clio-core",
+  "version": "0.1.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Run this script on the root directory.\" && exit 1"
+  },
+  "author": "Garrit Franke <garritfranke@gmail.com>",
+  "license": "Apache-2.0"
+}
