+++ packages/core/patterns.js
