+++ packages/core/tests/compiler.test.js
