+++ packages/core/tests/generator.test.js
