+++ packages/core/tests/lexer.test.js
