+++ packages/core/tests/parser.test.js
